Refactor types Add sanitizers
