+++ README.md
@@ -1,0 +1,4 @@
+TODO
+Use objectId for stores
+Better logging
+Publish common types
