+++ package.json
@@ -4,7 +4,7 @@
   "description": "Scrapes google shopping feeds and stores results in mongodb",
   "main": "dist/app.js",
   "type": "module",
-  "types": "types.d.ts",
+  "types": "types/index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "tsc -p ."
